Tools module collecting funcs from other modules.
