+++ spykeutils/spike_train_metrics.py
@@ -5,6 +5,7 @@ import quantities as pq
 import scipy as sp
 import _scipy_quantities as spq
 import signal_processing as sigproc
+import tools
 
 assert quantities_patch  # Suppress pyflakes warning, patch applied by loading
 
@@ -349,7 +350,7 @@ def st_inner(
 
 def _prepare_for_inner_prod(
         trains, smoothing_filter, sampling_rate, filter_area_fraction):
-    t_start, t_stop = sigproc.maximum_spike_train_interval({0: trains})
+    t_start, t_stop = tools.maximum_spike_train_interval({0: trains})
     padding = smoothing_filter.boundary_enclosing_at_least(filter_area_fraction)
     t_start -= 2 * padding
     t_stop += 2 * padding
