+++ spykeutils/tests/test_signal_processing.py
@@ -244,48 +244,6 @@ class TestRectangularKernel(ut.TestCase):
             actual.rescale(self.kernel_size.units), self.kernel_size)
 
 
-class Test_bin_spike_trains(ut.TestCase):
-    def test_bins_spike_train_using_its_properties(self):
-        a = neo.SpikeTrain(
-            sp.array([1.0]) * pq.s, t_start=0.5 * pq.s, t_stop=1.5 * pq.s)
-        sampling_rate = 4.0 * pq.Hz
-        expected = {0: [sp.array([0, 0, 1, 0])]}
-        expectedBins = sp.array([0.5, 0.75, 1.0, 1.25, 1.5]) * pq.s
-        actual, actualBins = sigproc.bin_spike_trains({0: [a]}, sampling_rate)
-        self.assertEqual(len(expected), len(actual))
-        self.assertEqual(len(expected[0]), len(actual[0]))
-        assert_array_equal(expected[0][0], actual[0][0])
-        assert_array_almost_equal(
-            expectedBins, actualBins.rescale(expectedBins.units))
-
-    def test_bins_spike_train_using_passed_properties(self):
-        a = neo.SpikeTrain(
-            sp.array([1.0]) * pq.s, t_start=0.0 * pq.s, t_stop=5.0 * pq.s)
-        sampling_rate = 4.0 * pq.Hz
-        t_start = 0.5 * pq.s
-        t_stop = 1.5 * pq.s
-        expected = {0: [sp.array([0, 0, 1, 0])]}
-        expectedBins = sp.array([0.5, 0.75, 1.0, 1.25, 1.5]) * pq.s
-        actual, actualBins = sigproc.bin_spike_trains(
-            {0: [a]}, sampling_rate=sampling_rate, t_start=t_start,
-            t_stop=t_stop)
-        self.assertEqual(len(expected), len(actual))
-        self.assertEqual(len(expected[0]), len(actual[0]))
-        assert_array_equal(expected[0][0], actual[0][0])
-        assert_array_almost_equal(
-            expectedBins, actualBins.rescale(expectedBins.units))
-
-    def test_uses_max_spike_train_interval(self):
-        a = arange_spikes(5 * pq.s)
-        b = arange_spikes(7 * pq.s, 15 * pq.s)
-        sampling_rate = 4.0 * pq.Hz
-        expectedBins = sp.arange(0.0, 15.1, 0.25) * pq.s
-        actual, actualBins = sigproc.bin_spike_trains(
-            {0: [a, b]}, sampling_rate=sampling_rate)
-        assert_array_almost_equal(
-            expectedBins, actualBins.rescale(expectedBins.units))
-
-
 class Test_smooth(ut.TestCase):
     def test_convolution_with_empty_binned_array_returns_array_of_zeros(self):
         binned = sp.zeros(10)
